STM32: Put some serial code in common between families
